added a separate TestUnionVector
